+++ src/Schema/Postgres/Compilers/CreateCompiler.php
@@ -16,14 +16,23 @@ class CreateCompiler
         array $columns,
         array $commands = []
     ): string {
+        if ($commands['like']) {
+            $postCompile = self::compileLike($grammar, $commands['like']);
+        } elseif ($commands['fromSelect']) {
+            $postCompile = self::compileFromSelect($grammar, $commands['fromSelect']);
+        } elseif ($commands['fromTable']) {
+            $postCompile = self::compileFromTable($grammar, $commands['fromTable']);
+        } else {
+            $postCompile = self::compileColumns($columns);
+        }
+
+
         $compiledCommand = sprintf(
-            '%s table %s %s (%s)',
-            $blueprint->temporary ? 'create temporary' : 'create',
+            'create%s table%s %s %s',
+            $blueprint->temporary ? ' temporary' : '',
             self::beforeTable($commands['ifNotExists']),
             $grammar->wrapTable($blueprint),
-            $commands['like']
-                ? self::compileLike($grammar, $commands['like'])
-                : self::compileColumns($columns)
+            $postCompile
         );
 
         return str_replace('  ', ' ', trim($compiledCommand));
@@ -31,18 +40,32 @@ class CreateCompiler
 
     private static function beforeTable(?Fluent $command = null): string
     {
-        return $command ? 'if not exists' : '';
+        return $command ? ' if not exists' : '';
     }
 
     private static function compileLike(Grammar $grammar, Fluent $command): string
     {
         $table        = $command->get('table');
         $includingAll = $command->get('includingAll') ? ' including all' : '';
-        return "like {$grammar->wrapTable($table)}$includingAll";
+        return "(like {$grammar->wrapTable($table)}$includingAll)";
+    }
+
+    private static function compileFromSelect(Grammar $grammar, Fluent $command): string
+    {
+        $sql = $command->get('fromSelect');
+
+        return "as ($sql)";
+    }
+
+    private static function compileFromTable(Grammar $grammar, Fluent $command): string
+    {
+        $table = $command->get('fromTable');
+
+        return "as TABLE $table";
     }
 
     private static function compileColumns(array $columns): string
     {
-        return implode(', ', $columns);
+        return '(' . implode(', ', $columns) . ')';
     }
 }